requirements.txt
dotenv==0.9.9
fastapi==0.116.1
google-auth==2.40.3
google-genai==1.25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
uvicorn==0.35.0